Move util to pkg and refractor util structure
