+++ pkg/util/ternary.go
@@ -1,21 +1,6 @@
 package util
 
-import (
-	"fmt"
-	"github.com/sirupsen/logrus"
-)
-
-// Panic if error exists
-func Must(err error, message string) {
-	if err != nil {
-		panic(fmt.Errorf("%s: %s\n", message, err.Error()))
-	}
-}
-
-// Converts interface to string
-func String(i interface{}) string {
-	return fmt.Sprintf("%v", i)
-}
+import "github.com/sirupsen/logrus"
 
 // Kind of ternary operator, used for conditional initialization
 type If bool
@@ -28,6 +13,20 @@ func (c If) String(a, b string) string {
 }
 
 func (c If) Int(a, b int) int {
+	if c {
+		return a
+	}
+	return b
+}
+
+func (c If) Float32(a, b float32) float32 {
+	if c {
+		return a
+	}
+	return b
+}
+
+func (c If) Float64(a, b float64) float64 {
 	if c {
 		return a
 	}
